+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 ### Fixed
 
 - Fix potential rounding issues in `offset()` function, when used as setter
+- Fix to ensure Slider items have `max-width: 100%`
+- Fix Dropdowns in Dropnav and Navbar components are initialized too late causing CLS
 
 ## 3.16.22 (June 22, 2023)